mock_api/requirements.txt
requests~=2.32.3
fastapi~=0.111.1
uvicorn~=0.30.3
psycopg2-binary~=2.9.9
sqlalchemy~=2.0.31
alembic~=1.13.2
python-dotenv~=1.0.1
pydantic~=2.8.2
pyjwt~=2.8.0
passlib~=1.7.4
python-jose~=3.3.0
cryptography~=43.0.0
asyncpg~=0.29.0